OpenDCL 9.2.1.4 - Added experimental support for GstarCAD 2024. - Copyright dates changed to 2024. - Fixed/silenced some build warnings.
git-svn-id: https://svn.code.sf.net/p/opendcl/code/Trunk@401 07a8c159-f786-4120-a571-e88efff3e5ae

+++ Common/Controls/GridCtrl.cpp
@@ -532,7 +532,7 @@ void CGridCtrl::SetCellStyle( int nRow, int nCol, Grid::CellStyle nStyle, int im
 		}
 	case Grid::SwitchableIcons:
 		{
-			LVITEM lvi = { LVIF_STATE, nRow, nCol, (altImage << 12), LVIS_STATEIMAGEMASK };
+			LVITEM lvi = { LVIF_STATE, nRow, nCol, (UINT(altImage) << 12), LVIS_STATEIMAGEMASK };
 			SetItem( &lvi );
 			break;
 		}
